refactor(home): move calendar logic to utils
